evals: added rich printing
