+++ evals/evals_allower.py
@@ -1,16 +1,18 @@
 """
-Checks is_scientific is set correctly.
+Checks that the is_scientific flag is set correctly.
 """
 
 import asyncio
+
+from rich.console import Console
 
 from askademic.allower import allower_agent
 from askademic.prompts import USER_PROMPT_ALLOWER_TEMPLATE
 
 
 class AllowerTestCase:
-    def __init__(self, request: str, is_scientic_gt: bool):
-        self.request = request
+    def __init__(self, question: str, is_scientic_gt: bool):
+        self.question = question
         self.is_scientic_gt = is_scientic_gt
 
 
@@ -23,26 +25,36 @@ eval_cases = [
     AllowerTestCase("What is the meaning of life?", False),
 ]
 
+console = Console()
+
 
 async def run_evals():
 
     c_passed, c_failed = 0, 0
     for case in eval_cases:
 
-        print(f"Evaluating case: {case.request}")
+        print(f"Evaluating case: {case.question}")
         response = await allower_agent.run(
-            USER_PROMPT_ALLOWER_TEMPLATE.format(question=case.request)
+            USER_PROMPT_ALLOWER_TEMPLATE.format(question=case.question)
         )
 
         if response.output.is_scientific != case.is_scientic_gt:
-            print(f"Test failed for question: {case.request}")
+            print(f"Test failed for question: {case.question}")
             c_failed += 1
         else:
             c_passed += 1
 
-    print(f"Total cases: {len(eval_cases)}")
-    print(f"Passed: {c_passed}")
-    print(f"Failed: {c_failed}")
+    console.print(f"[bold cyan]Total cases: {len(eval_cases)}[/bold cyan]")
+    if c_failed > 0:
+        print(
+            f"[bold green]Passed: {c_passed}[/bold green]"
+            + ","
+            + f"[bold red]Failed: {c_failed}[/bold red]"
+        )
+    console.print(
+        f"[bold green]Passed: {c_passed}[/bold green]",
+        style="bold",
+    )
 
 
 def main():
