+++ evals/evals_article.py
@@ -6,6 +6,8 @@ searching first by title and then by link.
 import asyncio
 import re
 
+from rich.console import Console
+
 from askademic.article import article_agent
 from askademic.prompts import USER_PROMPT_ARTICLE_TEMPLATE
 
@@ -51,6 +53,8 @@ eval_cases = [
 # we want to match regardless
 LINK_PATTERN = r"https?://arxiv\.org/pdf/(\d{4}\.\d{5})"
 
+console = Console()
+
 
 async def run_evals():
 
@@ -84,9 +88,17 @@ async def run_evals():
         else:
             c_passed += 1
 
-    print(f"Total cases: {len(eval_cases)}")
-    print(f"Passed: {c_passed}")
-    print(f"Failed: {c_failed}")
+    console.print(f"[bold cyan]Total cases: {len(eval_cases)}[/bold cyan]")
+    if c_failed > 0:
+        print(
+            f"[bold green]Passed: {c_passed}[/bold green]"
+            + ","
+            + f"[bold red]Failed: {c_failed}[/bold red]"
+        )
+    console.print(
+        f"[bold green]Passed: {c_passed}[/bold green]",
+        style="bold",
+    )
 
 
 def main():
