+++ evals/evals_summary.py
@@ -4,6 +4,8 @@ Checks category fetched is correct.
 
 import asyncio
 
+from rich.console import Console
+
 from askademic.summarizer import summary_agent
 
 
@@ -18,6 +20,8 @@ eval_cases = [
     SummaryTestCase("Can you summarize the latest papers on AI?", "cs.AI"),
     SummaryTestCase("Tell me all about the recent work in probability?", "stat.TH"),
 ]
+
+console = Console()
 
 
 async def run_evals():
@@ -34,9 +38,17 @@ async def run_evals():
         else:
             c_passed += 1
 
-    print(f"Total cases: {len(eval_cases)}")
-    print(f"Passed: {c_passed}")
-    print(f"Failed: {c_failed}")
+    console.print(f"[bold cyan]Total cases: {len(eval_cases)}[/bold cyan]")
+    if c_failed > 0:
+        print(
+            f"[bold green]Passed: {c_passed}[/bold green]"
+            + ","
+            + f"[bold red]Failed: {c_failed}[/bold red]"
+        )
+    console.print(
+        f"[bold green]Passed: {c_passed}[/bold green]",
+        style="bold",
+    )
 
 
 def main():
